Implemented CNCVC in CLVC
CNContactsViewController (CNCVC) is now ‘integrated’ in ContactsListViewController (CLVC).

-Bugs
If two contacts have the same name then the first contact in the array of fetched contacts will be shown. Fixed in CLVC sort of, explain later, but not in is in RecentlyAddedViewController (RAVC).

in CLVC, if you select a contact that no longer exists in your list of contacts on your device, the app will crash.